backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pdfplumber